+++ Music_Streaming/Music Streaming/Controllers/ArtistsController.cs
@@ -9,6 +9,9 @@ using Music_Streaming.Context;
 using Music_Streaming.Models;
 using Music_Streaming.ViewModels;
 using Music_Streaming.Mappers;
+using Microsoft.AspNetCore.Identity;
+using Music_Streaming.Authentication;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Music_Streaming.Controllers
 {
@@ -17,12 +20,16 @@ namespace Music_Streaming.Controllers
     public class ArtistsController : ControllerBase
     {
         private readonly MusicContext _context;
+        private readonly UserManager<ApplicationUser> userManager;
+        private readonly RoleManager<IdentityRole> roleManager;
 
 
 
-        public ArtistsController(MusicContext context)
+        public ArtistsController(MusicContext context, UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
         {
             _context = context;
+            this.userManager = userManager;
+            this.roleManager = roleManager;
         }
 
         // GET: api/Artists
@@ -57,7 +64,8 @@ namespace Music_Streaming.Controllers
                 List<SongViewModel> songViewModels = new List<SongViewModel>();
                 foreach (var song in songs)
                 {
-                    songViewModels.Add(Mapper.SongToViewModel(song));
+                    var artistman = await _context.Artists.Where(x => x.Id == item.ArtistId).FirstAsync();
+                    songViewModels.Add(Mapper.SongToViewModel(song,artist));
                 }
 
                 albumViewModels.Add(Mapper.AlbumToViewModel(item,songViewModels,Mapper.ArtistToViewModel(artist)));
@@ -102,28 +110,56 @@ namespace Music_Streaming.Controllers
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
         [HttpPost]
-        public async Task<ActionResult<Artist>> PostArtist(Artist artist)
+        [Authorize]
+        public async Task<ActionResult<ArtistViewModel>> PostArtist(ArtistViewModel artist)
         {
-            _context.Artists.Add(artist);
+            var user = await userManager.FindByNameAsync(User.Identity.Name);
+            if (user == null)
+            {
+                return Unauthorized("It stops here");
+            }
+            var checkArtist = await _context.Artists.Where(x => x.UserId == user.Id).FirstAsync();
+            if(checkArtist != null)
+            {
+                return BadRequest("User already assigned to an artist");
+            }
+
+            Artist artist1 = new Artist
+            {
+                Name = artist.Name,
+                UserId = user.Id
+            };
+            _context.Artists.Add(artist1);
             await _context.SaveChangesAsync();
 
-            return CreatedAtAction("GetArtist", new { id = artist.Id }, artist);
+            return CreatedAtAction("GetArtist", new { id = artist1.Id }, artist1);
         }
 
         // DELETE: api/Artists/5
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<ActionResult<Artist>> DeleteArtist(long id)
         {
+            var user = await userManager.FindByNameAsync(User.Identity.Name);
+            if (user == null)
+            {
+                return Unauthorized(new Response { Status = "Fail", Message = "I don't know how you could possibly get this error message but you managed to do it" });
+            }
+
             var artist = await _context.Artists.FindAsync(id);
             if (artist == null)
             {
-                return NotFound();
+                return NotFound(new Response { Status = "Fail", Message = "Artist does not exist" });
             }
 
+            if(artist.UserId != user.Id)
+            {
+                return Unauthorized(new Response { Status = "Fail", Message = "You do not own this album" });
+            }
             _context.Artists.Remove(artist);
             await _context.SaveChangesAsync();
 
-            return Ok();
+            return Ok(new Response { Status = "Success", Message = $"The artist-{artist.Name}- has been removed" });
         }
 
         private bool ArtistExists(long id)
